Conpherence - override some device specific styles so menu dropdown always present
Summary:
Fixes T7684. I am not sure if this is simple or not in CSS terms but I just blasted a bunch of css classes to make sure they have the proper "device-desktop" values for the conpherence durable column.

One small visual oddity is the main menu dropdowns (for conpherence and the actual page) should have a hover state css, but we add device-tablet too early relative to the actual page width. This is no big deal probably.

Test Plan: made the screen table sized. used the dropdown menu with great success!

Reviewers: chad, epriestley

Reviewed By: epriestley

Subscribers: Korvin, epriestley

Maniphest Tasks: T7684

Differential Revision: https://secure.phabricator.com/D12914