Conpherence - make solo conversations be titled with viewer's name
Summary: Fixes T8251. I think this used to be "lucky handle" when it was all one giant terrible function, so restore previous functionality and have the title be the viewer's name in this case.

Test Plan: rocked a solo conpherence and saw my username as the title

Reviewers: chad, epriestley

Reviewed By: epriestley

Subscribers: Korvin, epriestley

Maniphest Tasks: T8251

Differential Revision: https://secure.phabricator.com/D12935