Conpherence - make sure "Join" action in public room works even with no message text
Summary: Fixes T8328. Somewhere along the line we stopped posting to the server with no text. Make sure if the action is join_room that we ping the server even if no text is specified.

Test Plan: tried to send an empty message and failed; nothing happened when I clicked. tried to join a room with an empty message and it worked

Reviewers: epriestley

Reviewed By: epriestley

Subscribers: Korvin, epriestley

Maniphest Tasks: T8328

Differential Revision: https://secure.phabricator.com/D13041